moved input validation and bounding box management into new processdispatcher class
